refactor: Improve file hash storing to enable manual check of checksum
file-verification slot will now contain only the raw hash of the file, instead
of a json object. This will allow comparing hashes manually more easily

transclusion-hashes will now be set to empty string in case no transclusions
are found, instead of an empty array. This is also done so that is does not
pollute content_hash of the page, in case no transclusions are found